+++ tasks/index.js
@@ -131,7 +131,7 @@ task("masterchef:withdraw", "MasterChef withdraw")
 task("bar:enter", "SushiBar enter")
 .addParam("amount", "Amount")
 .setAction(async function ({ amount }, { ethers: { getNamedSigner } }, runSuper) {
-  const sushi = await ethers.getContract("SushiToken")
+  const sushi = await ethers.getContract("TiramisuToken")
 
   const bar = await ethers.getContract("SushiBar")
 
@@ -143,7 +143,7 @@ task("bar:enter", "SushiBar enter")
 task("bar:leave", "SushiBar leave")
 .addParam("amount", "Amount")
 .setAction(async function ({ amount }, { ethers: { getNamedSigner } }, runSuper) {
-  const sushi = await ethers.getContract("SushiToken")
+  const sushi = await ethers.getContract("TiramisuToken")
 
   const bar = await ethers.getContract("SushiBar")